LPS-147053 Extract publish button
